example_project/requirements.txt
# Install django-email-queue in editable mode, it's requirements for local development
# and this example in editable mode
# NOTE: paths are realative to CWD and not this file parent folder
# Running pip install -r example_project/requiremenst.txt

# This example requirements that may not not be already included in the app requirements
django
django-environ
gunicorn
psycopg2
whitenoise
